Clean up warnings (#3)
* Clean up all -Wunused-variable

* Clean up all -Wconversion

* Clean up all -Winteger-division

* Clean up all -Wunused-function

* Clean up all -Wcharacter-truncation

* Clean up all -Wtabs

* Clean up all -Wunused-dummy-argument

+++ src/lschpb.f90
@@ -136,7 +136,7 @@ subroutine lschpb(nn,ll,ierr,ee,rr,vv,uu,up,mmax,mch)
 ! start inward integration at 10*classical turning
 ! point with simple exponential
 
-         nin=mch+2.3d0/al
+         nin=int(mch+2.3d0/al)
          if(nin+4>mmax) nin=mmax-4
          xkap=dsqrt(sls/rr(nin)**2 + 2.0d0*(vv(nin)-ee))
